requirements.txt
boto3>=1.26.0
azure-identity>=1.7.0
azure-mgmt-resource>=20.0.0
azure-mgmt-authorization>=0.14.0
azure-monitor-query>=1.1.0
azure-mgmt-network>=20.0.0
azure-mgmt-storage>=20.0.0
